camps_ingest/requirements.txt
requests
beautifulsoup4
pdfplumber
pandas
python-slugify
phonenumbers
email-validator
tenacity
psycopg2-binary
python-dotenv